frontend-backend-demo/backend-flask/flask/src/requirements.txt
flask==2.0.3
python-dotenv==0.14.0
uwsgi==2.0.19.1
requests==2.28.1
